updated tests to run on geth
