+++ package.json
@@ -18,14 +18,14 @@
   "license": "ISC",
   "author": "eng@fragments.org",
   "scripts": {
-    "blockchain:start": "scripts/blockchain/runner.sh 1",
-    "blockchain:stop": "scripts/blockchain/runner.sh 0",
-    "precommit": "scripts/pre-commit.sh",
-    "genDummyData": "node_modules/truffle/build/cli.bundled.js --network gethUnitTest exec scripts/gen_dummy_data.js",
+    "blockchain:start": "./scripts/blockchain/runner.sh 1",
+    "blockchain:stop": "./scripts/blockchain/runner.sh 0",
+    "precommit": "./scripts/pre-commit.sh",
+    "genDummyData": "node_modules/truffle/build/cli.bundled.js --network gethUnitTest exec ./scripts/gen_dummy_data.js",
     "lint": "npm run lint:js && npm run lint:sol",
     "lint:js": "node_modules/eslint/bin/eslint.js ./ --cache --fix",
     "lint:sol": "node_modules/solhint/solhint.js contracts/*.sol",
-    "test": "scripts/test.sh ganacheUnitTest",
+    "test": "./scripts/test.sh",
     "trackGasUtilization": "npm run blockchain:start && npm run truffle exec test/load/gas_utilization.js save",
     "truffle": "node_modules/truffle/build/cli.bundled.js --network ganacheUnitTest"
   },
