+++ scripts/test.sh
@@ -9,50 +9,80 @@ source $DIR/blockchain/index.sh
 
 run-unit-tests(){
   frg-truffle \
-    --network $NETWORK_REF \
+    --network $1 \
+    --timeout 25000 \
     test \
-    $PROJECT_DIR/test/unit/*
+    $PROJECT_DIR/test/unit/*.js
 }
 
 run-simulation-tests(){
   frg-truffle \
-    --network $NETWORK_REF \
+    --network $1 \
     test \
     test/simulation/*
 }
 
 run-load-tests(){
   frg-truffle \
-    --network $NETWORK_REF \
+    --network $1 \
     exec \
     $PROJECT_DIR/test/load/gas_utilization.js verify
 }
 
-NETWORK_REF=$1
-read NETWORK_REF PORT < <(get-network-config $NETWORK_REF)
+read REF GANACHE_PORT < <(get-network-config ganacheUnitTest)
+read REF GETH_PORT < <(get-network-config gethUnitTest)
 
 # Is chain running?
-if [ $(process-pid $PORT) ]; then
-  REFRESH=0
+if [ $(process-pid $GANACHE_PORT) ]; then
+  REFRESH_GANACHE=0
 else
-  REFRESH=1
+  REFRESH_GANACHE=1
+fi
+if [ $(process-pid $GETH_PORT) ]; then
+  REFRESH_GETH=0
+else
+  REFRESH_GETH=1
 fi
 
-# Start chain
-start-chain $NETWORK_REF
-deploy-contracts $NETWORK_REF
-
-# Run tests
-run-unit-tests
-run-load-tests
+echo "------Start blockchain(s)"
+start-chain ganacheUnitTest
+deploy-contracts ganacheUnitTest
 if [ "${TRAVIS_EVENT_TYPE}" == "cron" ]; then
-  run-simulation-tests
+  start-chain gethUnitTest
+  deploy-contracts gethUnitTest
+fi
+
+echo "------Running gas utilization test"
+run-load-tests ganacheUnitTest
+
+echo "------Running unit tests"
+run-unit-tests ganacheUnitTest
+if [ "${TRAVIS_EVENT_TYPE}" == "cron" ]; then
+  run-unit-tests gethUnitTest
+fi
+
+if [ "${TRAVIS_EVENT_TYPE}" == "cron" ]; then
+  echo "------Running simulation tests"
+  run-simulation-tests ganacheUnitTest
+  run-simulation-tests gethUnitTest
 fi
 
 # Stop chain
-cleanup(){
-  if [ "$REFRESH" == "1" ]; then
-    stop-chain $NETWORK_REF
+cleanupGanache(){
+  if [ "$REFRESH_GANACHE" == "1" ]; then
+    stop-chain ganacheUnitTest
   fi
+  exit 0
 }
-trap cleanup EXIT
+
+cleanupGeth(){
+  if [ "$REFRESH_GETH" == "1" ]; then
+    stop-chain gethUnitTest
+  fi
+  exit 0
+}
+
+trap cleanupGanache EXIT
+if [ "${TRAVIS_EVENT_TYPE}" == "cron" ]; then
+  trap cleanupGeth EXIT
+fi
