+++ test/unit/erc20_behavior.js
@@ -1,22 +1,486 @@
-const uFragments = artifacts.require('UFragments.sol');
+const UFragments = artifacts.require('UFragments.sol');
 
 const _require = require('app-root-path').require;
-const { StandardTokenBehavior } = _require('/test/behavior/erc20');
+const BlockchainCaller = _require('/util/blockchain_caller');
+const chain = new BlockchainCaller(web3);
 
-contract('uFragments', function (accounts) {
-  let mFragments;
+const ZERO_ADDRESS = '0x0000000000000000000000000000000000000000';
+const transferAmount = 10;
+const erroneousAmount = 1001;
 
-  beforeEach('setup contract for each test', async function () {
-    mFragments = await uFragments.deployed();
+contract('UFragments:ERC20', function (accounts) {
+  let token;
+  before('setup UFragments contract', async function () {
+    token = await UFragments.new();
+  });
+  const initialTotalSupply = 1000;
+  const owner = accounts[0];
+  const anotherAccount = accounts[8];
+  const recipient = accounts[9];
+
+  describe('totalSupply', function () {
+    it('returns the total amount of tokens', async function () {
+      const totalSupply = await token.totalSupply.call();
+      assert.equal(totalSupply.toNumber(), initialTotalSupply);
+    });
   });
 
-  describe('uFragments as ERC20', function () {
-    it('should confirm to the specification', function () {
-      const owner = accounts[0];
-      const anotherAccount = accounts[8];
-      const recipient = accounts[9];
+  describe('balanceOf', function () {
+    describe('when the requested account has no tokens', function () {
+      it('returns zero', async function () {
+        const balance = await token.balanceOf.call(anotherAccount);
+        assert.equal(balance.toNumber(), 0);
+      });
+    });
+    describe('when the requested account has some tokens', function () {
+      it('returns the total amount of tokens', async function () {
+        const balance = await token.balanceOf.call(owner);
+        assert.equal(balance.toNumber(), initialTotalSupply);
+      });
+    });
+  });
 
-      StandardTokenBehavior(mFragments, 1000, owner, anotherAccount, recipient);
+  describe('transfer', function () {
+    describe('when the recipient is not the zero address', function () {
+      const to = recipient;
+
+      describe('when the sender does not have enough balance', function () {
+        const amount = erroneousAmount;
+
+        it('reverts', async function () {
+          await chain.expectEthException(token.transfer(to, amount, { from: owner }));
+        }).timeout();
+      });
+
+      describe('when the sender has enough balance', function () {
+        const amount = transferAmount;
+
+        afterEach(async () => {
+          // Ensuring state is consistent after the testing the specification
+          await token.transfer(owner, amount, { from: to });
+        });
+
+        it('transfers the requested amount & emits a transfer event', async function () {
+          const { logs } = await token.transfer(to, amount, { from: owner });
+
+          const senderBalance = await token.balanceOf.call(owner);
+          const supply = await token.totalSupply.call();
+          assert(supply.minus(amount).eq(senderBalance));
+
+          const recipientBalance = await token.balanceOf.call(to);
+          assert.equal(recipientBalance.toNumber(), amount);
+
+          assert.equal(logs.length, 1);
+          assert.equal(logs[0].event, 'Transfer');
+          assert.equal(logs[0].args.from, owner);
+          assert.equal(logs[0].args.to, to);
+          assert(logs[0].args.value.eq(amount));
+        });
+      });
+    });
+
+    describe('when the recipient is the zero address', function () {
+      const to = ZERO_ADDRESS;
+
+      it('reverts', async function () {
+        await chain.expectEthException(token.transfer(to, 10, { from: owner }));
+      });
+    });
+  });
+
+  describe('transfer from', function () {
+    const spender = recipient;
+
+    describe('when the recipient is not the zero address', function () {
+      const to = anotherAccount;
+
+      describe('when the spender has enough approved balance', function () {
+        beforeEach(async function () {
+          await token.approve(spender, 10, { from: owner });
+        });
+
+        describe('when the owner has enough balance', function () {
+          const amount = transferAmount;
+
+          afterEach(async () => {
+            // Ensuring state is consistent after the testing the specification
+            await token.transfer(owner, amount, { from: to });
+          });
+
+          it('transfers the requested amount, decreases the spender allowance, emits a transfer event', async function () {
+            const _senderBalance = await token.balanceOf.call(owner);
+            const { logs } = await token.transferFrom(owner, to, amount, { from: spender });
+            const senderBalance = await token.balanceOf.call(owner);
+            assert(_senderBalance.minus(amount).eq(senderBalance));
+            const recipientBalance = await token.balanceOf.call(to);
+            assert.equal(recipientBalance.toNumber(), amount);
+            const allowance = await token.allowance(owner, spender);
+            assert(allowance.eq(0));
+            assert.equal(logs.length, 1);
+            assert.equal(logs[0].event, 'Transfer');
+            assert.equal(logs[0].args.from, owner);
+            assert.equal(logs[0].args.to, to);
+            assert(logs[0].args.value.eq(amount));
+          });
+        });
+
+        describe('when the owner does not have enough balance', function () {
+          const amount = erroneousAmount;
+
+          it('reverts', async function () {
+            await chain.expectEthException(token.transferFrom(owner, to, amount, { from: spender }));
+          });
+        });
+      });
+
+      describe('when the spender does not have enough approved balance', function () {
+        beforeEach(async function () {
+          await token.approve(spender, 9, { from: owner });
+        });
+
+        describe('when the owner has enough balance', function () {
+          const amount = transferAmount;
+
+          it('reverts', async function () {
+            await chain.expectEthException(token.transferFrom(owner, to, amount, { from: spender }));
+          });
+        });
+
+        describe('when the owner does not have enough balance', function () {
+          const amount = erroneousAmount;
+
+          it('reverts', async function () {
+            await chain.expectEthException(token.transferFrom(owner, to, amount, { from: spender }));
+          });
+        });
+      });
+    });
+
+    describe('when the recipient is the zero address', function () {
+      const amount = transferAmount;
+      const to = ZERO_ADDRESS;
+
+      beforeEach(async function () {
+        await token.approve(spender, amount, { from: owner });
+      });
+
+      it('reverts', async function () {
+        await chain.expectEthException(token.transferFrom(owner, to, amount, { from: spender }));
+      });
+    });
+  });
+
+  describe('approve', function () {
+    describe('when the spender is not the zero address', function () {
+      const spender = recipient;
+
+      describe('when the sender has enough balance', function () {
+        const amount = transferAmount;
+
+        it('emits an approval event', async function () {
+          const { logs } = await token.approve(spender, amount, { from: owner });
+
+          assert.equal(logs.length, 1);
+          assert.equal(logs[0].event, 'Approval');
+          assert.equal(logs[0].args.owner, owner);
+          assert.equal(logs[0].args.spender, spender);
+          assert(logs[0].args.value.eq(amount));
+        });
+
+        describe('when there was no approved amount before', function () {
+          it('approves the requested amount', async function () {
+            await token.approve(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, amount);
+          });
+        });
+
+        describe('when the spender had an approved amount', function () {
+          beforeEach(async function () {
+            await token.approve(spender, 1, { from: owner });
+          });
+
+          it('approves the requested amount and replaces the previous one', async function () {
+            await token.approve(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, amount);
+          });
+        });
+      });
+
+      describe('when the sender does not have enough balance', function () {
+        const amount = erroneousAmount;
+
+        it('emits an approval event', async function () {
+          const { logs } = await token.approve(spender, amount, { from: owner });
+
+          assert.equal(logs.length, 1);
+          assert.equal(logs[0].event, 'Approval');
+          assert.equal(logs[0].args.owner, owner);
+          assert.equal(logs[0].args.spender, spender);
+          assert(logs[0].args.value.eq(amount));
+        });
+
+        describe('when there was no approved amount before', function () {
+          it('approves the requested amount', async function () {
+            await token.approve(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, amount);
+          });
+        });
+
+        describe('when the spender had an approved amount', function () {
+          beforeEach(async function () {
+            await token.approve(spender, 1, { from: owner });
+          });
+
+          it('approves the requested amount and replaces the previous one', async function () {
+            await token.approve(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, amount);
+          });
+        });
+      });
+    });
+
+    describe('when the spender is the zero address', function () {
+      const amount = transferAmount;
+      const spender = ZERO_ADDRESS;
+
+      it('approves the requested amount', async function () {
+        await token.approve(spender, amount, { from: owner });
+
+        const allowance = await token.allowance(owner, spender);
+        assert.equal(allowance, amount);
+      });
+
+      it('emits an approval event', async function () {
+        const { logs } = await token.approve(spender, amount, { from: owner });
+
+        assert.equal(logs.length, 1);
+        assert.equal(logs[0].event, 'Approval');
+        assert.equal(logs[0].args.owner, owner);
+        assert.equal(logs[0].args.spender, spender);
+        assert(logs[0].args.value.eq(amount));
+      });
+    });
+  });
+
+  describe('decrease approval', function () {
+    describe('when the spender is not the zero address', function () {
+      const spender = recipient;
+
+      beforeEach(async () => {
+        await token.approve(spender, 0, { from: owner });
+      });
+
+      describe('when the sender has enough balance', function () {
+        const amount = transferAmount;
+
+        it('emits an approval event', async function () {
+          const { logs } = await token.decreaseApproval(spender, amount, { from: owner });
+
+          assert.equal(logs.length, 1);
+          assert.equal(logs[0].event, 'Approval');
+          assert.equal(logs[0].args.owner, owner);
+          assert.equal(logs[0].args.spender, spender);
+          assert(logs[0].args.value.eq(0));
+        });
+
+        describe('when there was no approved amount before', function () {
+          it('keeps the allowance to zero', async function () {
+            await token.decreaseApproval(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, 0);
+          });
+        });
+
+        describe('when the spender had an approved amount', function () {
+          beforeEach(async function () {
+            await token.approve(spender, amount + 1, { from: owner });
+          });
+
+          it('decreases the spender allowance subtracting the requested amount', async function () {
+            await token.decreaseApproval(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, 1);
+          });
+        });
+      });
+
+      describe('when the sender does not have enough balance', function () {
+        const amount = erroneousAmount;
+
+        it('emits an approval event', async function () {
+          const { logs } = await token.decreaseApproval(spender, amount, { from: owner });
+
+          assert.equal(logs.length, 1);
+          assert.equal(logs[0].event, 'Approval');
+          assert.equal(logs[0].args.owner, owner);
+          assert.equal(logs[0].args.spender, spender);
+          assert(logs[0].args.value.eq(0));
+        });
+
+        describe('when there was no approved amount before', function () {
+          it('keeps the allowance to zero', async function () {
+            await token.decreaseApproval(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, 0);
+          });
+        });
+
+        describe('when the spender had an approved amount', function () {
+          beforeEach(async function () {
+            await token.approve(spender, amount + 1, { from: owner });
+          });
+
+          it('decreases the spender allowance subtracting the requested amount', async function () {
+            await token.decreaseApproval(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, 1);
+          });
+        });
+      });
+    });
+
+    describe('when the spender is the zero address', function () {
+      const amount = transferAmount;
+      const spender = ZERO_ADDRESS;
+
+      beforeEach(async () => {
+        await token.approve(spender, 0, { from: owner });
+      });
+
+      it('decreases the requested amount', async function () {
+        await token.decreaseApproval(spender, amount, { from: owner });
+
+        const allowance = await token.allowance(owner, spender);
+        assert.equal(allowance, 0);
+      });
+
+      it('emits an approval event', async function () {
+        const { logs } = await token.decreaseApproval(spender, amount, { from: owner });
+
+        assert.equal(logs.length, 1);
+        assert.equal(logs[0].event, 'Approval');
+        assert.equal(logs[0].args.owner, owner);
+        assert.equal(logs[0].args.spender, spender);
+        assert(logs[0].args.value.eq(0));
+      });
+    });
+  });
+
+  describe('increase approval', function () {
+    const amount = transferAmount;
+
+    describe('when the spender is not the zero address', function () {
+      const spender = recipient;
+
+      beforeEach(async () => {
+        await token.approve(spender, 0, { from: owner });
+      });
+
+      describe('when the sender has enough balance', function () {
+        it('emits an approval event', async function () {
+          const { logs } = await token.increaseApproval(spender, amount, { from: owner });
+
+          assert.equal(logs.length, 1);
+          assert.equal(logs[0].event, 'Approval');
+          assert.equal(logs[0].args.owner, owner);
+          assert.equal(logs[0].args.spender, spender);
+          assert(logs[0].args.value.eq(amount));
+        });
+
+        describe('when there was no approved amount before', function () {
+          it('approves the requested amount', async function () {
+            await token.increaseApproval(spender, amount, { from: owner });
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance.toNumber(), amount);
+          });
+        });
+
+        describe('when the spender had an approved amount', function () {
+          beforeEach(async function () {
+            await token.approve(spender, 1, { from: owner });
+          });
+
+          it('increases the spender allowance adding the requested amount', async function () {
+            await token.increaseApproval(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, amount + 1);
+          });
+        });
+      });
+
+      describe('when the sender does not have enough balance', function () {
+        const amount = erroneousAmount;
+
+        it('emits an approval event', async function () {
+          const { logs } = await token.increaseApproval(spender, amount, { from: owner });
+
+          assert.equal(logs.length, 1);
+          assert.equal(logs[0].event, 'Approval');
+          assert.equal(logs[0].args.owner, owner);
+          assert.equal(logs[0].args.spender, spender);
+          assert(logs[0].args.value.eq(amount));
+        });
+
+        describe('when there was no approved amount before', function () {
+          it('approves the requested amount', async function () {
+            await token.increaseApproval(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance.toNumber(), amount);
+          });
+        });
+
+        describe('when the spender had an approved amount', function () {
+          beforeEach(async function () {
+            await token.approve(spender, 1, { from: owner });
+          });
+
+          it('increases the spender allowance adding the requested amount', async function () {
+            await token.increaseApproval(spender, amount, { from: owner });
+
+            const allowance = await token.allowance(owner, spender);
+            assert.equal(allowance, amount + 1);
+          });
+        });
+      });
+    });
+
+    describe('when the spender is the zero address', function () {
+      const spender = ZERO_ADDRESS;
+
+      beforeEach(async () => {
+        await token.approve(spender, 0, { from: owner });
+      });
+
+      it('approves the requested amount', async function () {
+        await token.increaseApproval(spender, amount, { from: owner });
+
+        const allowance = await token.allowance(owner, spender);
+        assert.equal(allowance, amount);
+      });
+
+      it('emits an approval event', async function () {
+        const { logs } = await token.increaseApproval(spender, amount, { from: owner });
+
+        assert.equal(logs.length, 1);
+        assert.equal(logs[0].event, 'Approval');
+        assert.equal(logs[0].args.owner, owner);
+        assert.equal(logs[0].args.spender, spender);
+        assert(logs[0].args.value.eq(amount));
+      });
     });
   });
 });
